Fix event broadcasts with reusable archives. Fix passing unique_ptr arguments.
Use of smart_ptrs for polymorphic events collides with
the object tracking performed by Cereal. This prevents
reuse of the binary archive. unique_ptrs are not tracked,
so they are now used to pass the event objects.
To simplify the Service class interface polymorphic event objects
are now created internally by the broadcast_event function template.